+++ templates/app/edge/template/app/app.component.js
@@ -1,0 +1,6 @@
+import template from './app.component.html';
+import './app.component.scss';
+
+export default {
+  template
+};