+++ templates/component/directive-view/v1/template.js
@@ -1,13 +1,13 @@
 var path = require('path');
 var chalk = require('chalk');
-var Template = require('../../../lib/template');
+var Template = require('../../../../lib/template');
 var Case = require('case');
 
 module.exports = Template.extend({
 
   init: function() {
     this._super.apply(this, arguments);
-    this.description = 'An anglar.directive component for fine-grained control.'
+    this.description = 'An angular.directive component with UI Router state.'
   }
 
 });