+++ templates/component/directive/v2/template/name.controller.js
@@ -1,0 +1,21 @@
+import styles from './<%= name %>.scss';
+
+export default class {
+
+  constructor() {
+    'ngInject';
+
+    //css-modules (https://github.com/webpack/css-loader#css-modules)
+    this.styles = styles;
+  }
+
+  $onInit() {
+  }
+
+  $onChanges() {
+  }
+
+  $onDestroy() {
+  }
+  
+}