+++ templates/component/directive/v2/template/name.module.js
@@ -1,8 +1,8 @@
 import <%=name_camel%>Styles from './<%= name %>.scss';
-import <%=name_camel%>Component from './<%= name %>.component';
+import <%=name_camel%>Directive from './<%= name %>.directive';
 
 var module = angular.module('<%= moduleName %>', []);
 
-module.component('<%= name_camel %>', <%=name_camel%>Component);
+module.directive('<%= name_camel %>', <%=name_camel%>Directive);
 
 export default module.name;