+++ templates/component/view/v2/template/name.component.js
@@ -1,0 +1,9 @@
+import template from './<%= name %>.html';
+import controller from './<%= name %>.controller';
+
+export default {
+  template,
+  controller,
+  bindings: {
+  }
+};